First Attempt for Class Generation on DB Project
