+++ Contractor.Core/Contractor.Core.csproj
@@ -49,10 +49,10 @@
     <None Update="Projects\Contract.Persistence\Templates\IEntitiesCrudRepositoryTemplate.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Projects\DB\Templates\TableTemplate-RequestScope.txt">
+    <None Update="Projects\DB\Templates\EntitiesTableTemplate-RequestScope.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-    <None Update="Projects\DB\Templates\TableTemplate.txt">
+    <None Update="Projects\DB\Templates\EntitiesTableTemplate.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
     <None Update="Projects\Logic.Tests\Templates\EntityCreateTestTemplate.txt">
