+++ Contractor.Core/Projects/DB/Generation/DbTable/Services/DbTablePropertyAddition.cs
@@ -3,7 +3,7 @@ using Contractor.Core.Options;
 using Contractor.Core.Tools;
 using System.IO;
 
-namespace Contractor.Core.Projects
+namespace Contractor.Core.Projects.DB
 {
     internal class DbTablePropertyAddition
     {
