+++ Contractor.Core/Projects/DB/Templates/EntitiesTableTemplate-RequestScope.txt
