+++ Contractor.Core/Projects/DB/Templates/EntitiesTableTemplate.txt
